+++ compiler/Grammar/Antlr/CircularQueue.hpp
@@ -53,7 +53,7 @@ public:
 		{ if (storage.size()>0) storage.erase(storage.begin()); }
 	void append(const T& t)
 		{ storage.push_back(t); }
-	int entries() const
+	size_t entries() const
 		{ return storage.size(); }
 
 private:
